refactor: chunk up into smaller components
