+++ src/pages/User/contact/UserContactForm.tsx
@@ -51,10 +51,7 @@ export const UserContactForm = observer(({ user }: Props) => {
   }
 
   return (
-    <Flex
-      sx={{ flexDirection: 'column', marginY: 2 }}
-      data-cy="UserContactForm"
-    >
+    <Flex sx={{ flexDirection: 'column' }} data-cy="UserContactForm">
       <Heading as="h3" variant="small" my={2}>
         {`${title} ${user.displayName}`}
       </Heading>
