+++ src/routes/_.u.$id._index.tsx
@@ -1,6 +1,6 @@
 import { json } from '@remix-run/node'
 import { useLoaderData } from '@remix-run/react'
-import { UserProfile } from 'src/pages/User/content/UserProfile'
+import { ProfilePage } from 'src/pages/User/content/ProfilePage'
 import { pageViewService } from 'src/services/pageViewService.server'
 import { userService } from 'src/services/userService.server'
 import { generateTags, mergeMeta } from 'src/utils/seo.utils'
@@ -63,5 +63,5 @@ export default function Index() {
     )
   }
 
-  return <UserProfile profile={profile} userCreatedDocs={userCreatedDocs} />
+  return <ProfilePage profile={profile} userCreatedDocs={userCreatedDocs} />
 }
